Rename imports/ui/games to reduce confusion once inprogress games are added
